// src/models/IBudgetRow.ts
export interface IBudgetRow {
  id?: number;
  parentId?: number | null;
  child?: IBudgetRow[];
  equipmentCosts: number;
  estimatedProfit: number;
  machineOperatorSalary: number;
  mainCosts: number;
  materials: number;
  mimExploitation: number;
  overheads: number;
  rowName: string;
  salary: number;
  supportCosts: number;
  total: number;
}